+++ React/react-pro/src/components/ui/Home.jsx
@@ -1,6 +1,6 @@
-import LandingPage from "../../pages/ui-pg/LandingPage";
-import Contact from "./Contact";
-import Footer from "./Footer";
+import LandingPage from "../../pages/home-pg/LandingPage";
+import Contact from "../../pages/home-pg/Contact";
+import Footer from "../../pages/home-pg/Footer";
 import React from "react";
 
 const Home = () => {
