+++ React/react-pro/src/pages/home-pg/home-pg.css
@@ -1,0 +1,191 @@
+/* Landing Page */
+
+.landing-pg {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	width: 100%;
+	height: 100vh;
+	background-color: #000;
+	background-image: linear-gradient(62deg, #3a3d40 0%, #181719 100%);
+}
+
+.landing-pg > p {
+	font-size: 3rem;
+	font-weight: 200;
+	font-style: italic;
+	color: var(--main-red);
+}
+
+/* Contact section */
+.contact-section {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	text-align: center;
+	width: 100%;
+	height: 80vh;
+	padding: 0 2rem;
+	background: var(--main-gray);
+}
+
+.contact-section-header > h2 {
+	font-size: 6rem;
+}
+
+@media (max-width: 28.75em) {
+	.contact-section-header > h2 {
+		font-size: 4rem;
+	}
+}
+
+.contact-section-header > p {
+	font-style: italic;
+}
+
+.contact-links {
+	display: flex;
+	justify-content: center;
+	width: 100%;
+	max-width: 980px;
+	margin-top: 4rem;
+	flex-wrap: wrap;
+}
+
+.contact-icons {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	flex-direction: column;
+	position: relative;
+	transform: translate(-50%, -50%);
+	width: fit-content;
+	height: auto;
+}
+
+/* styled-icons */
+
+.logo-icon {
+	width: 200px;
+	height: 70px;
+	background: linear-gradient(to left top, #2d0087 50%, #480101 50%);
+	border-style: none;
+	color: rgb(0, 0, 0);
+	font-size: 23px;
+	letter-spacing: 3px;
+	font-family: "Lato";
+	font-weight: 600;
+	outline: none;
+	cursor: pointer;
+	position: relative;
+	padding: 0px;
+	overflow: hidden;
+	transition: all 0.5s;
+	box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.2);
+}
+
+.icon span {
+	position: absolute;
+	display: block;
+}
+.icon span:nth-child(1) {
+	height: 3px;
+	width: 200px;
+	top: 0px;
+	left: -200px;
+	background: linear-gradient(to right, rgba(0, 0, 0, 0), #f6e58d);
+	border-top-right-radius: 1px;
+	border-bottom-right-radius: 1px;
+	animation: span1 2s linear infinite;
+	animation-delay: 1s;
+}
+
+@keyframes span1 {
+	0% {
+		left: -200px;
+	}
+	100% {
+		left: 200px;
+	}
+}
+.icon span:nth-child(2) {
+	height: 70px;
+	width: 3px;
+	top: -70px;
+	right: 0px;
+	background: linear-gradient(to bottom, rgba(0, 0, 0, 0), #f6e58d);
+	border-bottom-left-radius: 1px;
+	border-bottom-right-radius: 1px;
+	animation: span2 2s linear infinite;
+	animation-delay: 2s;
+}
+@keyframes span2 {
+	0% {
+		top: -70px;
+	}
+	100% {
+		top: 70px;
+	}
+}
+.icon span:nth-child(3) {
+	height: 3px;
+	width: 200px;
+	right: -200px;
+	bottom: 0px;
+	background: linear-gradient(to left, rgba(0, 0, 0, 0), #f6e58d);
+	border-top-left-radius: 1px;
+	border-bottom-left-radius: 1px;
+	animation: span3 2s linear infinite;
+	animation-delay: 3s;
+}
+@keyframes span3 {
+	0% {
+		right: -200px;
+	}
+	100% {
+		right: 200px;
+	}
+}
+
+.icon span:nth-child(4) {
+	height: 70px;
+	width: 3px;
+	bottom: -70px;
+	left: 0px;
+	background: linear-gradient(to top, rgba(0, 0, 0, 0), #f6e58d);
+	border-top-right-radius: 1px;
+	border-top-left-radius: 1px;
+	animation: span4 2s linear infinite;
+	animation-delay: 4s;
+}
+@keyframes span4 {
+	0% {
+		bottom: -70px;
+	}
+	100% {
+		bottom: 70px;
+	}
+}
+
+.icon:hover {
+	transition: all 0.5s;
+	transform: rotate(-3deg) scale(1.1);
+	box-shadow: 0px 3px 5px rgba(0, 0, 0, 0.4);
+}
+.icon:hover span {
+	animation-play-state: paused;
+}
+
+/*
+a {
+	color: rgb(152, 12, 180);
+}
+
+a .icon {
+	color: rgb(25, 5, 29);
+	border-radius: 50px;
+	background: linear-gradient(145deg, #f0f0f0, #cacaca);
+	box-shadow: 20px 20px 60px #bebebe, -20px -20px 60px #ffffff;
+} */
