+++ React/react-yarn/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "react-yarn",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "react-scripts": "^5.0.1"
+  }
+}
